Title: millionaire question
A millionaire met a stranger who talked to him about a plan to exchange money. The plan is as follows: I give you 100,000 yuan a day, you give me 1 cent the first day, 2 cents the second day, 4 cents the third day... After 30 days of this exchange, how much did the millionaire pay? How much did the stranger hand over? (note: one is ten thousand yuan, one is the point) input description: the output of two integers, respectively represents the millionaire to hand over the money and strangers to hand over the money, the millionaire to hand over the money in ten thousand yuan, strangers to hand over the money as a unit of points.